Resumable Earley parsing!
Parser now remembers the tokens it parsed with last time, and will resume from
an earlier point table if possible. Completer now supports both parsing and
completing, which share a single leftParser object.

Still to do: memory management for shared parsing. At the moment the result
will use the same parse structures as last time, which are based around the
same Token objects. This is a Bad Thing, since currently we rely on that to do
syntax highlighting.

I propose moving AST creation *outside* the parser. This should be slightly
more efficient, since then we won't create partial ASTs for parse fragments
which are, in the end, unused. Note that Earley is currently used for three
things:
- the completer. This only produces completion objects.
- syntax highlighting. If we attach colours to Rules, perhaps we can eliminate AST
  creation & the paint() phase.
- the final compile phase: this obviously needs the AST.

Then we can replace every Parser and Completer with a *shared* Completer
object, and call .parse() and .complete() on that. Then things should
(hopefully) be more efficient :-)